fix: ensure deployscript sets directory contract
Something that came up during testing:
The deployment script didn't set the directory contract in both, the
voting contract and the featured voting contract, which resulted in
functoin calls to revert because they essentially didn't exist at the
addresses referenced in those contracts.

This commit fixes it by ensuring `setDirectory()` is called for both
voting contracts.