fix(VotingContract): voting rooms are considered active until evaluated
This commit fixes a bug in the `VotingContract` where any voting room
marked as `finalized` is not considered to be an active voting room,
even though, evaluation is still ongoing.

The fix adds a check that a voting room has to be either not marked as
finalized, or, if it is, it should not be evaluated, to be considered
an active voting room.

It also adds a test that ensures the fix is working as expected.